house_ingest/pyproject.toml
[tool]
[tool.poetry]
name = "house_ingest"
version = "0.1.0"
homepage = "https://github.com/briggySmalls/house_ingest"
description = "Top-level package for house ingest."
authors = ["Sam Briggs <briggySmalls90@gmail.com>"]
readme = "README.rst"
license =  "MIT"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.5',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
]
packages = [
    { include = "house_ingest" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.8,<4"
click = "*"
rightmove-webscraper = {git = "git@github.com:briggySmalls/rightmove_webscraper.py.git"}
planparser = {git = "git@github.com:briggySmalls/planparser.git"}
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
elasticsearch = "^8.4.3"
tqdm = "^4.64.1"

[tool.poetry.dev-dependencies]
bumpversion = "*"
coverage = "*"
flake8 = "*"
invoke = "*"
isort = "*"
pylint = "*"
pytest = ">=7.2.0"
sphinx = ">=5.3.0"
black = "*"

[tool.poetry.scripts]
houseingest = 'house_ingest.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
